Back off specific import, move LinearDictionary
