+++ api-report/merge-tree.api.md
@@ -867,6 +867,9 @@ export class List<T> {
 }
 
 // @public (undocumented)
+export function ListMakeHead<U>(): List<U>;
+
+// @public (undocumented)
 export function loadSegments(content: string, segLimit: number, markers?: boolean, withProps?: boolean): ISegment[];
 
 // @public (undocumented)
