+++ packages/dds/merge-tree/src/test/beastTest.ts
@@ -13,9 +13,13 @@ import { DebugLogger } from "@fluidframework/telemetry-utils";
 import { ISequencedDocumentMessage } from "@fluidframework/protocol-definitions";
 import JsDiff from "diff";
 import random from "random-js";
-import { Property } from "../base";
 import {
-    LinearDictionary,
+    KeyComparer,
+    Property,
+    PropertyAction,
+    SortedDictionary,
+} from "../base";
+import {
     ProxString,
     RedBlackTree,
     Stack,
@@ -48,6 +52,97 @@ import {
 import { specToSegment, TestClient } from "./testClient";
 import { TestServer } from "./testServer";
 import { insertText, loadTextFromFile, nodeOrdinalsHaveIntegrity } from "./testUtils";
+
+function LinearDictionary<TKey, TData>(compareKeys: KeyComparer<TKey>): SortedDictionary<TKey, TData> {
+    const props: Property<TKey, TData>[] = [];
+    const compareProps = (a: Property<TKey, TData>, b: Property<TKey, TData>) => compareKeys(a.key, b.key);
+    function diag() {
+        console.log(`size is ${props.length}`);
+    }
+    function mapRange<TAccum>(action: PropertyAction<TKey, TData>, accum?: TAccum, start?: TKey, end?: TKey) {
+        let _start = start;
+        let _end = end;
+
+        if (props.length !== 0) { return; }
+
+        if (_start === undefined) {
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+            _start = min()!.key;
+        }
+        if (_end === undefined) {
+            // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+            _end = max()!.key;
+        }
+        for (let i = 0, len = props.length; i < len; i++) {
+            if (compareKeys(_start, props[i].key) <= 0) {
+                const ecmp = compareKeys(_end, props[i].key);
+                if (ecmp < 0) {
+                    break;
+                }
+                if (!action(props[i], accum)) {
+                    break;
+                }
+            }
+        }
+    }
+
+    function map<TAccum>(action: PropertyAction<TKey, TData>, accum?: TAccum) {
+        mapRange(action, accum);
+    }
+
+    function min() {
+        if (props.length > 0) {
+            return props[0];
+        }
+    }
+    function max() {
+        if (props.length > 0) {
+            return props[props.length - 1];
+        }
+    }
+
+    function get(key: TKey) {
+        for (let i = 0, len = props.length; i < len; i++) {
+            if (props[i].key == key) {
+                return props[i];
+            }
+        }
+    }
+
+    function put(key: TKey, data: TData) {
+        if (key !== undefined) {
+            if (data === undefined) {
+                remove(key);
+            }
+            else {
+                props.push({ key, data });
+                props.sort(compareProps); // Go to insertion sort if too slow
+            }
+        }
+    }
+    function remove(key: TKey) {
+        if (key !== undefined) {
+            for (let i = 0, len = props.length; i < len; i++) {
+                if (props[i].key == key) {
+                    props[i] = props[len - 1];
+                    props.length--;
+                    props.sort(compareProps);
+                    break;
+                }
+            }
+        }
+    }
+    return {
+        min,
+        max,
+        map,
+        mapRange,
+        remove,
+        get,
+        put,
+        diag,
+    };
+}
 
 let logLines: string[];
 function log(message: any) {
